Move deployment attributes to own class
