+++ server/src/main/java/org/jboss/as/server/deployment/DeploymentHandlerUtils.java
@@ -25,9 +25,6 @@ import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.ARC
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.HASH;
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.PATH;
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.RELATIVE_TO;
-import static org.jboss.as.server.controller.resources.DeploymentResourceDescription.CONTENT_BYTES;
-import static org.jboss.as.server.controller.resources.DeploymentResourceDescription.CONTENT_INPUT_STREAM_INDEX;
-import static org.jboss.as.server.controller.resources.DeploymentResourceDescription.CONTENT_URL;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
@@ -39,7 +36,7 @@ import java.util.List;
 import org.jboss.as.controller.OperationContext;
 import org.jboss.as.controller.OperationFailedException;
 import org.jboss.as.server.ServerMessages;
-import org.jboss.as.server.controller.resources.DeploymentResourceDescription;
+import org.jboss.as.server.controller.resources.DeploymentAttributes;
 import org.jboss.dmr.ModelNode;
 
 /**
@@ -75,21 +72,21 @@ abstract class DeploymentHandlerUtils {
 
     protected static InputStream getInputStream(OperationContext context, ModelNode operation) throws OperationFailedException {
         InputStream in = null;
-        if (operation.hasDefined(CONTENT_INPUT_STREAM_INDEX.getName())) {
-            int streamIndex = CONTENT_INPUT_STREAM_INDEX.resolveModelAttribute(context, operation).asInt();
+        if (operation.hasDefined(DeploymentAttributes.CONTENT_INPUT_STREAM_INDEX.getName())) {
+            int streamIndex = DeploymentAttributes.CONTENT_INPUT_STREAM_INDEX.resolveModelAttribute(context, operation).asInt();
             int maxIndex = context.getAttachmentStreamCount();
             if (streamIndex > maxIndex) {
-                throw ServerMessages.MESSAGES.invalidStreamIndex(CONTENT_INPUT_STREAM_INDEX.getName(), streamIndex, maxIndex);
+                throw ServerMessages.MESSAGES.invalidStreamIndex(DeploymentAttributes.CONTENT_INPUT_STREAM_INDEX.getName(), streamIndex, maxIndex);
             }
             in = context.getAttachmentStream(streamIndex);
-        } else if (operation.hasDefined(CONTENT_BYTES.getName())) {
+        } else if (operation.hasDefined(DeploymentAttributes.CONTENT_BYTES.getName())) {
             try {
-                in = new ByteArrayInputStream(CONTENT_BYTES.resolveModelAttribute(context, operation).asBytes());
+                in = new ByteArrayInputStream(DeploymentAttributes.CONTENT_BYTES.resolveModelAttribute(context, operation).asBytes());
             } catch (IllegalArgumentException iae) {
-                throw ServerMessages.MESSAGES.invalidStreamBytes(CONTENT_BYTES.getName());
+                throw ServerMessages.MESSAGES.invalidStreamBytes(DeploymentAttributes.CONTENT_BYTES.getName());
             }
-        } else if (operation.hasDefined(CONTENT_URL.getName())) {
-            final String urlSpec = CONTENT_URL.resolveModelAttribute(context, operation).asString();
+        } else if (operation.hasDefined(DeploymentAttributes.CONTENT_URL.getName())) {
+            final String urlSpec = DeploymentAttributes.CONTENT_URL.resolveModelAttribute(context, operation).asString();
             try {
                 in = new URL(urlSpec).openStream();
             } catch (MalformedURLException e) {
@@ -113,7 +110,7 @@ abstract class DeploymentHandlerUtils {
      * @return {@code true} of the parameter is valid, otherwise {@code false}.
      */
     protected static boolean hasValidContentAdditionParameterDefined(ModelNode operation) {
-        for (String s : DeploymentResourceDescription.MANAGED_CONTENT_ATTRIBUTES.keySet()) {
+        for (String s : DeploymentAttributes.MANAGED_CONTENT_ATTRIBUTES.keySet()) {
             if (operation.hasDefined(s)) {
                 return true;
             }
